+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 ### Fixed
 
+* Fixed [#598](https://github.com/NeoVintageous/NeoVintageous/issues/598): Text object registers should be linewise
 * Fixed [#597](https://github.com/NeoVintageous/NeoVintageous/issues/597): `di(` and other text objects ignore indention
 * Fixed [#591](https://github.com/NeoVintageous/NeoVintageous/issues/591): `?` (search forward) in visual modes is inconsistent
 * Fixed [#590](https://github.com/NeoVintageous/NeoVintageous/issues/590): `/` (search backward) in visual modes is inconsistent